cohere-connector/pyproject.toml
[tool.poetry]
name = "pinecone-search"
version = "0.1.0"
description = ""
authors = ["Bach Tran <bachtran@seas.upenn.edu>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
flask = "2.2.5"
connexion = {extras = ["swagger-ui"], version = "^2.14.2"}
pinecone-client = "^2.2.4"
python-dotenv = "^1.0.0"
black = "^23.11.0"
cohere = "4.44"
gunicorn = "^21.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
